Finalizado consumo da API
