+++ src/views/News.vue
@@ -3,7 +3,7 @@
     <v-app-bar app shrink-on-scroll>
       <!-- <v-app-bar-nav-icon></v-app-bar-nav-icon> -->
 
-      <v-toolbar-title>Pesquisa de Notícias</v-toolbar-title>
+      <v-toolbar-title>Feed de Notícias</v-toolbar-title>
 
       <v-spacer></v-spacer>
 
@@ -11,32 +11,99 @@
         <v-icon>mdi-dots-vertical</v-icon>
       </v-btn> -->
     </v-app-bar>
-
-    <v-main>
-      <v-container>
-        <news></news>
-        <v-row>
-          <v-col v-for="n in 15" :key="n" cols="4">
-            <v-card height="200"></v-card>
-          </v-col>
-        </v-row>
-      </v-container>
-    </v-main>
+    <div class="news">
+      <p>Pesquisa de Notícias</p>
+      <form class="form" v-on:submit.prevent="searchNews">
+        <input
+          type="text"
+          class="assunto-input"
+          v-model="assunto"
+          placeholder="Assunto"
+          maxlength="8"
+        />
+        <input type="submit" value="Pesquisar" class="assunto-button" />
+      </form>
+      <CardNews :data="data"></CardNews>
+    </div>
   </v-app>
 </template>
 
 <script>
-import News from "../components/News.vue";
+import axios from "axios"; // biblioteca usada para fazer requisições HTTP, similar ao Fetch
+import CardNews from "../components/CardNews";
 
 export default {
   name: "News",
   data() {
-    return {};
+    return {
+      assunto: "",
+      data: {
+        count: 0,
+        items: [
+          {
+            titulo: "",
+            introducao: "",
+            data_publicacao: "",
+            link: "",
+          },
+        ],
+      },
+      status: Number,
+      isError: false,
+    };
+  },
+  methods: {
+    searchNews: function() {
+      if (this.assunto.length > 0) {
+        axios
+          .get(
+            `http://servicodados.ibge.gov.br/api/v3/noticias/?busca=${this.assunto}/json`
+          )
+          .then((response) => {
+            this.status = response.status;
+            this.data = response.data;
+            console.log(response);
+          })
+          .catch(
+            // caso a Promise seja rejeitada cairá no catch
+            (error) => {
+              console.log(error);
+              this.isError = true;
+            }
+          );
+      }
+    },
+  },
+  computed: {
+    tituloUpper() {
+      return this.data.items.titulo.toUpperCase();
+    },
   },
   components: {
-    news: News,
+    CardNews,
   },
-  methods: {},
-  computed: {},
 };
 </script>
+
+<style scoped>
+.assunto-input {
+  border: 1px solid black;
+  border-radius: 5px;
+  padding: 5px;
+  width: 250px;
+  margin: 10px;
+}
+
+.assunto-button {
+  border: 1px solid black;
+  border-radius: 5px;
+  width: 100px;
+  background-color: teal;
+  color: white;
+  margin: 10px;
+}
+
+.city-button:hover {
+  background-color: rgb(155, 155, 155);
+}
+</style>
